задание 2.14 тест

+++ src/main/java/ru/ssau/tk/lenuska0506/doska/Task22.java
@@ -59,7 +59,7 @@ class Task214 {
 
     static void setarr(double[] arr) {
         for (int i = 0; i < arr.length; i++) {
-            arr[i] = -((Double )arr[i]) ;
+            arr[i] = -(arr[i]);
         }
     }
 }
